+++ misc/examples/smartpointers/person_arc.c
@@ -13,7 +13,7 @@ int Person_cmp(const Person* a, const Person* b) {
     return c ? c : cstr_cmp(&a->last, &b->last);
 }
 
-uint64_t Person_hash(const Person* a) {
+size_t Person_hash(const Person* a) {
     return cstr_hash(&a->name) ^ cstr_hash(&a->last);
 }
